# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: NexBank Digital Banking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with filters
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPageResponse"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Platform statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user details with accounts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /admin/users/{id}/suspend:
    post:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend user account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User suspended
  /admin/users/{id}/unsuspend:
    post:
      operationId: unsuspendUser
      tags: [admin]
      summary: Unsuspend user account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unsuspended
  /admin/fund-account:
    post:
      operationId: fundAccount
      tags: [admin]
      summary: Credit user account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundAccountInput"
      responses:
        "200":
          description: Account funded
  /admin/debit-account:
    post:
      operationId: debitAccount
      tags: [admin]
      summary: Debit user account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebitAccountInput"
      responses:
        "200":
          description: Account debited
  /admin/transactions:
    get:
      operationId: listAdminTransactions
      tags: [admin]
      summary: All platform transactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /admin/transactions/{id}/reverse:
    post:
      operationId: reverseTransaction
      tags: [admin]
      summary: Reverse a transaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction reversed
  /admin/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [admin]
      summary: Audit log
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          {}
      required:
        - success

    # ── Auth Schemas ──────────────────────────────────────────────────
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        userId:
          type: integer
        refNumber:
          type: string
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, TRANSFER_IN, TRANSFER_OUT, FEE, INTEREST, REVERSAL]
        amount:
          type: string
        balanceBefore:
          type: string
        balanceAfter:
          type: string
        currency:
          type: string
        description:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, REVERSED]
        counterpartName:
          type: ["string", "null"]
        counterpartBank:
          type: ["string", "null"]
        counterpartAccountNumber:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, refNumber, type, amount, status]
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...
    TransactionPageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Transaction"
            pagination:
              $ref: "#/components/schemas/Pagination"
...
    FundAccountInput:
      type: object
      required: [accountId, amount]
      properties:
        accountId:
          type: integer
        amount:
          type: string
        description:
          type: ["string", "null"]
    DebitAccountInput:
      type: object
      required: [accountId, amount]
      properties:
        accountId:
          type: integer
        amount:
          type: string
        description:
          type: ["string", "null"]